Foodsaveconnect-master/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
pillow>=10.0
